editor/libeditor/tests/test_bug767684.html

<!DOCTYPE html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=767684
-->
<title>Test for Bug 767684</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=767684">Mozilla Bug 767684</a>
<div contenteditable>foo<b>bar</b>baz</div>
<script>
getSelection().selectAllChildren(document.querySelector("div"));
document.execCommand("increaseFontSize");
is(document.querySelector("div").innerHTML, "<big>foo<b>bar</b>baz</big>",
   "All selected text must be embiggened");
</script>
